Moved get_type to _compat.py to satisfy linter. Updated conftest.py to ignore 563 tests for py36 and py37. Attempted to configure tox.ini to customize linting by python version.
