+++ src/cattr/_compat.py
@@ -17,7 +17,7 @@ unicode = str
 bytes = bytes
 
 if not is_py36:
-    from typing import List, Union, _GenericAlias
+    from typing import List, Union, _GenericAlias, ForwardRef, Any
 
     def is_union_type(obj):
         return (
@@ -71,6 +71,20 @@ if not is_py36:
             or args == bare_mutable_seq_args
         )
 
+    @lru_cache(32)
+    def _get_class_namespace(class_: type) -> Dict[str, Any]:
+        return vars(sys.modules[class_.__module__])
+
+    @lru_cache(32)
+    def get_type(parent_class: type, type_: Union[type, str]) -> type:
+        """Returns the actual type if type_ is a string."""
+
+        if isinstance(type_, str):
+            globalns = _get_class_namespace(parent_class)
+            forwardRef = ForwardRef(type_, is_argument=False)
+            return forwardRef._evaluate(globalns, None)
+        else:
+            return type_
 
 else:
     from typing import _Union
@@ -99,3 +113,6 @@ else:
         return not type.__args__
 
     is_bare_frozenset = is_bare
+
+    def get_type(ignored: type, type_: type) -> type:
+        return type_
