+++ tox.ini
@@ -1,9 +1,21 @@
 [tox]
-envlist = pypy3, py36, py37, py38, py39, lint
+envlist = pypy3, py36, py37, py38, py39, lint-{py36,py37}
 
-[testenv:lint]
+[testenv:lint-py36]
 skip_install = true
 basepython = python3.6
+extras = tests
+deps =
+    flake8
+    black
+commands =
+    flake8 --exclude=pep563*,test_563.py src tests setup.py conftest.py docs/conf.py
+    black --exclude=pep563*,test_563.py --check --verbose setup.py src tests docs/conf.py
+
+[testenv:lint-py37]
+skip_install = true
+basepython = 
+    py37: python3.7
 extras = tests
 deps =
     flake8
